Issue #812: handle delete in no extension case if ocr is enabled

+++ core/app/tika/tika-common/src/main/java/io/openk9/tika/TikaProcessor.java
@@ -261,7 +261,9 @@ public class TikaProcessor {
                 else {
                     if (!retainBinaries) {
 //                        logger.info("Deleting resource with id: " + resourceId + " and name: " + name);
-                        fileManagerClient.delete(resourceId, schemaName);
+                        if (!ocrEnabled) {
+                            fileManagerClient.delete(resourceId, schemaName);
+                        }
                         logger.info("Skipping resource with id: " + resourceId + " and name: " + name
                         + " because not supported by configuration");
                     }
